CalyxOS KEEP: first_run: deactivate autoupdate globally
We currently do not have any update service infrastructure in place
(i.e. on our webserver), and sending update requests to Google also
sounds meh (wrong provider, after all).

Also, there is no tunable in chrome://settings (or similar) yet to
turn it back on.

CalyxOS Note: This was removed in Cromite 120 as part of "Introduce
Chrome web store protection", but we are not desktop, so we do not
support extensions and do not include that patch. So we keep this.

License: GPL-3.0-only - https://spdx.org/licenses/GPL-3.0-only.html

Change-Id: I10bb5eaf40d22a112c4a16a4987e8799b1b00094