profile-resetter: disable send-settings
When the user wishes to reset his preferences, Chromium would suggest
to send over the old settings to Google. I presume that this is for
Google to investigate into the setting combinations that led the user
to issue a rest.

Of course, we cannot let that happen by default. Google is not to get
any information.

License: GPL-3.0-only - https://spdx.org/licenses/GPL-3.0-only.html

Change-Id: I77d20f8b4a09c4de1a9e36d719970e5e52e44a8d