Disable sending URLs to Google for metrics
Disable the "Make searches and browsing better" feature, which
claims to send URLs of visited pages to Google. If the toggle is not
removed from the UI via a separate patch, it may appear possible to
turn this on, but the change will not be committed.

Taken from the "Disable safe browsing" Bromite patch, which is
originally based on ungoogled-chromium patch.

License: GPL-3.0-only - https://spdx.org/licenses/GPL-3.0-only.html
Change-Id: I75e381891a09b6621ab5aef79e85022655513bb8