Remove history-saving option for "Add an always-incognito mode"
Ultimately I think the option makes sense, but currently it is not
properly restricted to scenarios in which always-on incognito is
active. Even if it is prevented from being turned on except when
always-incognito is on, there is a window between that point and
relaunch in which it affects incognito behavior, and I strongly
believe it should *only* affect incognito for always-on incognito
to avoid breaking user expectations. That problem would need to be
resolved before bringing this back.

Change-Id: I446731c5b6c90519cdd71bbc4d0fbdd1db7e2ab0