Disable additional segmentation platform features
Fix a crash on non-first startup related to home module ranking,
which relies on segmentation platform which is already disabled via
Remove-segmentation-platform.patch. A Chromium fix would be ideal
in concept so that when the core feature (kSegmentationPlatformFeature)
is disabled, its "subfeatures" are not still consulted -- at least,
that would make sense to me. However, it's easy for us to just turn off
these "subfeatures" ourselves for now.

Change-Id: Id37707403644bb83f4ac0df49d7baf86b8b94b81