Temp workaround for home modules crashing
These are still leading to crashes, even following the change entitled
"Disable additional segmentation platform features." For now, we shall
strip out the crashing code -- which we have been expecting *shouldn't*
run, anyway, given that we disabled their feature flags -- from the
individual crash sites until there is more time to figure out a better
way to fix this.

Notably, the "Your last tab" card still works sometimes. I'm not
entirely sure what triggers it to appear for us, but it must not be
totally broken as a result of this change and the aforementioned one,
because I have seen it on multiple occasions. Perhaps that card isn't
even a "home module". More research is needed.

Change-Id: Ie0aadd2cc507c355c1e93066521e4294f867538b