Don't save payment methods or addresses by default
Changing the default states directly affects existing users, too.
When such settings are on by default, it is non-trivial to determine
whether the user wants to keep it that way, apart from asking them
or coming up with some type of heuristic, which would only be a guess.
We currently have no facility to ask *or* to notify of a change, and
we do not want a jarring change to occur without consent. Therefore,
we only want to adjust these defaults for new users.

Solidify the defaults for saving payment methods and addresses:
- Alter the actual defaults to turn these off.
- For existing users who did not make a choice, these would have been on
  by default. To maintain this behavior, turn them on explicitly, as if
  user-chosen.
- On first startup, set the new default values explicitly, as if they
  were user-chosen. This prevents them from being overwritten on
  subsequent startups via the above behavior.

Issue: calyxos#3060
Issue: calyxos#2758
Change-Id: I31a55afbe6e9d5a8cab3f2b32e45c101c6751f70